Continuing networking
Need to send request for full information based on bookRowID

+++ MyApplicationProject/MyApplication/src/main/java/com/example/myapplication/endless_list_activity.java
@@ -54,8 +54,8 @@ public class endless_list_activity extends Activity implements FragmentEndlessLi
     }
 
     @Override
-    public void onItemSelected(int id) {
-        Log.e("Item Clicked: ",Integer.toString(id));
+    public void onItemSelected(long id) {
+        Log.e("Item Clicked: ",Long.toString(id));
         Intent detailIntent = new Intent(this, ItemDetailActivity.class);
         detailIntent.putExtra("BOOK_ROW_ID",id);
         startActivity(detailIntent);
